refactor(core): extract shared windowsPath utility and fix PATH override semantics
Address PR #3451 review comments:

1. [Critical] Fix PATH override semantics: normalize process.env before
   merging with server config so that a server-provided PATH fully
   replaces the parent value instead of being merged with a stale
   case-variant (Path).

2. [Suggestion] Extract mergeWindowsPathValues and
   normalizePathEnvForWindows into a shared utility module
   (src/utils/windowsPath.ts), eliminating near-identical copies in
   shellExecutionService.ts and mcp-client.ts. The shared version
   includes the fingerprint caching from shellExecutionService.

3. [Suggestion] Revert the "should connect via command" test to its
   original non-Windows form, restoring generic coverage. Add a
   dedicated test for server config PATH override behavior.
