+++ packages/core/src/tools/mcp-client.test.ts
@@ -272,12 +272,6 @@ describe('mcp-client', () => {
     });
 
     it('should connect via command', async () => {
-      vi.spyOn(process, 'platform', 'get').mockReturnValue('win32');
-      process.env = {
-        ...ORIGINAL_ENV,
-        PATH: 'C:\\Windows\\System32;C:\\Shared\\Tools',
-        Path: 'C:\\Users\\tester\\bin;C:\\Shared\\Tools',
-      };
       const mockedTransport = vi
         .spyOn(SdkClientStdioLib, 'StdioClientTransport')
         .mockReturnValue({} as SdkClientStdioLib.StdioClientTransport);
@@ -297,14 +291,9 @@ describe('mcp-client', () => {
         command: 'test-command',
         args: ['--foo', 'bar'],
         cwd: 'test/cwd',
-        env: expect.objectContaining({
-          FOO: 'bar',
-          PATH: 'C:\\Windows\\System32;C:\\Shared\\Tools;C:\\Users\\tester\\bin',
-        }),
+        env: { ...process.env, FOO: 'bar' },
         stderr: 'pipe',
       });
-      const transportOptions = mockedTransport.mock.calls[0]?.[0];
-      expect(transportOptions?.env?.['Path']).toBeUndefined();
     });
 
     it('should normalize PATH-like env keys on Windows for stdio transport', async () => {
@@ -338,6 +327,32 @@ describe('mcp-client', () => {
         stderr: 'pipe',
       });
       const transportOptions = mockedTransport.mock.calls[0]?.[0];
+      expect(transportOptions?.env?.['Path']).toBeUndefined();
+    });
+
+    it('should let server config PATH override parent PATH on Windows', async () => {
+      vi.spyOn(process, 'platform', 'get').mockReturnValue('win32');
+      process.env = {
+        ...ORIGINAL_ENV,
+        PATH: 'C:\\Windows\\System32;C:\\Shared\\Tools',
+        Path: 'C:\\Users\\tester\\bin;C:\\Shared\\Tools',
+      };
+      const mockedTransport = vi
+        .spyOn(SdkClientStdioLib, 'StdioClientTransport')
+        .mockReturnValue({} as SdkClientStdioLib.StdioClientTransport);
+
+      await createTransport(
+        'test-server',
+        {
+          command: 'test-command',
+          env: { PATH: 'C:\\ServerToolchain\\bin' },
+        },
+        false,
+      );
+
+      const transportOptions = mockedTransport.mock.calls[0]?.[0];
+      // Server-provided PATH should fully replace the parent PATH, not merge
+      expect(transportOptions?.env?.['PATH']).toBe('C:\\ServerToolchain\\bin');
       expect(transportOptions?.env?.['Path']).toBeUndefined();
     });
